requirements.txt
Flask #  latest
Flask-CORS #  latest
pandas#  latest
Pillow # latest (PIL fork)
asgiref # latest
python-dotenv #  latest
google-generativeai # latest